refactor classes to model
