+++ StrongholdCrusader/src/main/java/org/example/model/Player.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.model;
 
 public class Player {
